apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
openai==1.59.7
prisma==0.15.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
langchain==0.3.16
langchain-openai==0.2.14
langchain-community==0.3.16
